refactor: move src/http/routes.ts to src/adapters/http/
